pyproject.toml
[project]
name = "ai-alert"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "guardrails-ai>=0.8.0",
    "langchain>=1.2.9",
    "langchain-openai>=1.1.8",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "selenium>=4.40.0",
    "sqlalchemy>=2.0.46",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
